Implementando CSS do mobile da página SignUp
