--- src/App.tsx
@@ -1,12 +1,0 @@
-import './App.css';
-import SignUp from './pages/SignUp/signUp';
-
-function App() {
-  return (
-    <div>
-      <SignUp/>   
-    </div>   
-  );
-}
-
-export default App;
