--- src/components/ExemploComponent/ExemploComponent.tsx
@@ -1,1 +1,0 @@
-export{};
