+++ src/components/FormSignUpMobile/formSignUp.css
@@ -1,0 +1,99 @@
+body {
+    background-color: #ffffff;
+    min-width: 100vw;
+    min-height: 100vh;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+.create-account {
+    min-width: 100%;
+    min-height: 100%;
+    padding: 55px 90px;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center; 
+    
+    background-color: #E0F2F1;
+}
+
+.img-logo {
+    width: 99px;
+    height: 166px;
+
+    margin-top: 60px;
+    margin-bottom: 30px;
+}
+
+.form_group {
+    margin-bottom: 2rem;
+    position: relative;
+}
+
+.input, .group_uf {
+    width: 100%;
+    height: 45px;
+    font-size: 18px;
+    padding-left: 8px;
+    outline: none;
+    border-radius: 0.5rem;
+    border: 1px solid #ccc;  
+    background-color: #408D86;
+    color: #fff;
+}
+.group_uf option {
+    color: #fff;
+    font-weight: 400;
+}
+
+.form_sign-up {
+    width: 100%;
+    height: 45px;
+    border-radius: 0.5rem;
+    border: 1px solid #263339;
+    background-color: #263339;
+    color: #ffffff;
+    cursor: pointer;
+    transition: background-color .3s ease;
+}
+
+.form_sign-up:hover {
+    background-color: #408D86;
+    border: 1px solid #408D86;
+}
+
+.input-label {
+    position: absolute;
+    left: 5px;
+    top: 10px;
+    pointer-events: none;
+    transition: 0.5s;
+    color: #fff;
+}
+
+.create-account_form input:focus ~ label {
+    transform: translateY(-33px);
+    font-size: 16px;
+    opacity: 1;
+    color: #408D86;
+}
+
+.form_checkbox {
+    display: flex;
+}
+
+::-webkit-calendar-picker-indicator {
+    filter: invert(1);
+}
+
+input[type=date]{ color: transparent; }
+input[type=date]:valid{ color: #fff; }
+input[type=date]:focus{ color: #fff; }
+
+.confirm-terms-and-conditions {
+    margin-left: 5px;
+}
