+++ src/index.css
@@ -11,3 +11,25 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
+
+:root {
+  --backgroundmobile-color: #E0F2F1;
+  --darkbtn-color: #263339;
+  --backgrounddesktop-color: #80CBC4;
+  --hoverbotaoescuropraclaro--color: #80CBC4;
+  --hoverbotaoclaropraescuro--color: #408d86;
+  font-family: 'Roboto', sans-serif;
+}
+
+* {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+}
+
+#root {
+  width: 100%;
+  margin: 10px;
+}
