+++ src/index.tsx
@@ -1,13 +1,14 @@
-import React from 'react';
+import * as React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
+import SignUp from './pages/SignUp/signUp';
+
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    <App/>
+    <SignUp/>
   </React.StrictMode>
 );
