--- src/pages/ExemploPage/ExemploPage.tsx
@@ -1,1 +1,0 @@
-export {};
